SecretWarden use Active Objects & Exclusion Paths feature.
With Atlassian plugins, there are two key ways of storing plugin
settings. The first, what SecretWarden previously used - pluginSettings
is a basic key-value store, with a limited set of features. So limited,
it is not really compatible with Data Center - its doesn't replicate
across nodes for a start. This was solved by introducing caching, but
was far more difficult to add additional settings.

Then there is Active Objects. Active Objects allows you to store
arbitary data inside the data. There's really, three key parts to it.
The first is an entity. You specify some getters/setters in an interface
and extend from entity. This creates a table in the db with
AOxxxxx_TABLE. You have a DAO interface (Data Access Object) which provides a way
of accessing the stored data, and an implementation of this. Active
Objects have a lot of advantages including being transactional &
cluster-safe without caching. There is also something called
'UpgradeTasks' (not required at the moment) that lets you change the
schema of already stored data. This change converts from PluginSettings
to Active Objects.

This change also introduces a new feature to excluded paths (based of
Regular Expressions)